switch: arizona: Always set HPD_GND_SEL and MICDx_GND_SEL together
These values should apparently always be set to the same thing whilst
running a detection as we currently set them independantly during HPD and
MICD this will not be the case. Furthermore, since we can't be sure if a
HPDET or MICD will run first simply set both values during both operations.

Change-Id: I7f1d1be14d378a4e79f211a04431b67f4014bd0b
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>